fix(bayn): complete candidate precommit contract
